# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the Care-Wallet App.",
        "title": "Care-Wallet API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/group/{groupId}/labels": {
            "get": {
                "description": "get all labels for a group given their group id",
                "tags": [
                    "labels"
                ],
                "summary": "get labels for a group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the group id to get labels for",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Label"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "create a new label for a group",
                "tags": [
                    "labels"
                ],
                "summary": "Create A New Label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group to create label for",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label creation data",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/labels.LabelData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Label"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/group/{groupId}/labels/{lname}": {
            "delete": {
                "description": "delete a label",
                "tags": [
                    "labels"
                ],
                "summary": "Delete A Label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group to delete label from",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of label to delete",
                        "name": "lname",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "edit a label",
                "tags": [
                    "labels"
                ],
                "summary": "Edit A Label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Group of label to edit",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Name of label to edit",
                        "name": "lname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Label edit data",
                        "name": "_",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/labels.LabelData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Label"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/tasks": {
            "post": {
                "description": "Create a new task",
                "tags": [
                    "tasks"
                ],
                "summary": "Create a New Task",
                "parameters": [
                    {
                        "description": "Create Task Request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created Task",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                }
            }
        },
...
        "/tasks/{tid}": {
            "get": {
                "description": "get a task given its id",
                "tags": [
                    "tasks"
                ],
                "summary": "get task by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the id of the task",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the task_info field of a task by ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Update Task Info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Task Info Request",
                        "name": "request_body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasks.TaskBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated Task",
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task by ID",
                "tags": [
                    "tasks"
                ],
                "summary": "Delete a Task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Task ID",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/tasks/{tid}/labels": {
            "get": {
                "description": "get a tasks labels given the task id",
                "tags": [
                    "task labels"
                ],
                "summary": "get a tasks labels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "the task id to get labels for",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Task_Label"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "add a label to a task given the task id, group id, and label name",
                "tags": [
                    "task labels"
                ],
                "summary": "add a label to a task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the task id to add the label to",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The label data to add to the task",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task_labels.LabelData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Task_Label"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "remove a label from a task given the task id, group id, and label name",
                "tags": [
                    "task labels"
                ],
                "summary": "remove a label from a task",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the task id to get labels for",
                        "name": "tid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The label data to remove from the task",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task_labels.LabelData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "labels.LabelData": {
            "type": "object",
            "properties": {
                "label_color": {
                    "type": "string"
                },
                "label_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Label": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "label_color": {
                    "type": "string"
                },
                "label_name": {
                    "type": "string"
                }
            }
        },
...
        "models.Task": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "User ID",
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "quick_task": {
                    "type": "boolean"
                },
                "repeating": {
                    "type": "boolean"
                },
                "repeating_end_date": {
                    "type": "string"
                },
                "repeating_interval": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "task_id": {
                    "type": "integer"
                },
                "task_info": {
                    "type": "string"
                },
                "task_status": {
                    "type": "string"
                },
                "task_title": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                }
            }
        },
...
        "models.Task_Label": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "label_name": {
                    "type": "string"
                },
                "task_id": {
                    "type": "integer"
                }
            }
        },
...
        "task_labels.LabelData": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "label_name": {
                    "type": "string"
                }
            }
        },
...
        "tasks.TaskBody": {
            "type": "object",
            "properties": {
                "created_by": {
                    "description": "User ID",
                    "type": "string"
                },
                "created_date": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "notes": {
                    "type": "string"
                },
                "quick_task": {
                    "type": "boolean"
                },
                "repeating": {
                    "type": "boolean"
                },
                "repeating_end_date": {
                    "type": "string"
                },
                "repeating_interval": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "task_info": {
                    "type": "string"
                },
                "task_status": {
                    "type": "string"
                },
                "task_type": {
                    "type": "string"
                }
            }
        },
...
    }
}